// 13.heroesApp/src/app/heroes/interfaces/hero.interface.ts
// Generated by https://quicktype.io

export interface Hero {
    id:               string;
    superhero:        string;
    publisher:        Publisher;
    alter_ego:        string;
    first_appearance: string;
    characters:       string;
    alt_img?:         string; // creamos esta propiedad opcional
}
// un enum, solo son posibles esas dos opciones
export enum Publisher {
    DCComics = "DC Comics",
    MarvelComics = "Marvel Comics",
}